webapi/tct-webmessaging-w3c-tests/webmessaging-py/w3c/postMessage_asterisk_xorigin.htm

<!DOCTYPE html>
<html>
<head>
<title> Cross-origin postMessage with targetOrigin == "*" </title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/server.js?pipe=sub"></script>
</head>
<body>
<div id=log></div>

<div id=post style="display:none">
</div>

<script>
    setup({timeout:500});

    var description = "Test Description: To send the message to the target regardless of origin, set the target origin to '*'.";

    var t = async_test(description);

    var div = document.getElementById("post");
    var TARGET1 = document.createElement("iframe");
    var TARGET2 = document.createElement("iframe");
    div.appendChild(TARGET1);
    div.appendChild(TARGET2);

    var XORIGIN = "http://" + __SERVER__NAME + ":" + __CORS__PORT;
    var SORIGIN = "http://" + location.host;
    var ExpectedResult = ["#1", XORIGIN, "#2", SORIGIN];
    var ActualResult = [];

    TARGET2.onload = PostMessageTest;
    TARGET1.src = "http://" + __SERVER__NAME + ":" + __CORS__PORT + "/tests/webmessaging/w3c/support/ChildWindowPostMessage.htm";
    TARGET2.src = "./support/ChildWindowPostMessage.htm";

    function PostMessageTest()
    {
        TARGET1.contentWindow.postMessage("#1", "*");
        TARGET2.contentWindow.postMessage("#2", "*");
    }

    window.onmessage = t.step_func(function(e)
    {
        ActualResult.push(e.data, e.origin);

        if (ActualResult.length >= ExpectedResult.length)
        {
            assert_array_equals(ActualResult, ExpectedResult, "ActualResult");
            t.done();
        }
    });
</script>
</body>
</html>
